server_py/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydub==0.25.1
librosa==0.10.1
yt-dlp==2023.11.16
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.3
soundfile==0.12.1
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
pytube==15.0.0
